python/requirements.txt
certifi==2018.1.18
chardet==3.0.4
cognitive-face==1.4.1
idna==2.6
numpy==1.14.2
opencv-python==3.4.0.12
pyserial==3.4
requests==2.18.4
urllib3==1.22
